chore: centralize constant in one place
