+++ project/tests/users/views/test_account_view.py
@@ -5,9 +5,8 @@ from django.conf import settings
 from rest_framework import status
 from rest_framework.test import APIClient
 
-from users.constants import TokenScope, EventType
 from users.models import CustomUser, SecurityQuestion, UserQuestionAnswer
-from common.constants import ErrorTypes
+from common.constants import ErrorTypes, TokenScope, EventType
 from tests.helpers.setup_mock_accounts import (
     setup_mock_accounts,
     ACTIVE_USER_USERNAME,
